move files in lib to root project dir
